DTSCCI-682 Refactor EvidenceUploadHandlerBase in civil-service
