+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/EvidenceUploadHandlerBase.java
@@ -12,10 +12,11 @@ import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
 import uk.gov.hmcts.reform.civil.documentmanagement.model.Document;
 import uk.gov.hmcts.reform.civil.enums.CaseRole;
-import uk.gov.hmcts.reform.civil.enums.MultiPartyScenario;
 import uk.gov.hmcts.reform.civil.enums.caseprogression.EvidenceUploadExpert;
 import uk.gov.hmcts.reform.civil.enums.caseprogression.EvidenceUploadTrial;
 import uk.gov.hmcts.reform.civil.enums.caseprogression.EvidenceUploadWitness;
+import uk.gov.hmcts.reform.civil.handler.callback.user.task.evidenceuploadhandlerbasetask.DocumentUploadTimeTask;
+import uk.gov.hmcts.reform.civil.handler.callback.user.task.evidenceuploadhandlerbasetask.SetOptionsTask;
 import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
 import uk.gov.hmcts.reform.civil.model.Bundle;
 import uk.gov.hmcts.reform.civil.model.BusinessProcess;
@@ -24,7 +25,6 @@ import uk.gov.hmcts.reform.civil.model.IdValue;
 import uk.gov.hmcts.reform.civil.model.caseprogression.UploadEvidenceDocumentType;
 import uk.gov.hmcts.reform.civil.model.caseprogression.UploadEvidenceExpert;
 import uk.gov.hmcts.reform.civil.model.caseprogression.UploadEvidenceWitness;
-import uk.gov.hmcts.reform.civil.model.common.DynamicList;
 import uk.gov.hmcts.reform.civil.model.common.Element;
 import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
 import uk.gov.hmcts.reform.civil.service.CoreCaseUserService;
@@ -60,8 +60,6 @@ import static uk.gov.hmcts.reform.civil.callback.CallbackType.SUBMITTED;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.EVIDENCE_UPLOADED;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.EVIDENCE_UPLOAD_APPLICANT;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.EVIDENCE_UPLOAD_RESPONDENT;
-import static uk.gov.hmcts.reform.civil.enums.CaseCategory.SPEC_CLAIM;
-import static uk.gov.hmcts.reform.civil.enums.CaseCategory.UNSPEC_CLAIM;
 import static uk.gov.hmcts.reform.civil.enums.CaseRole.RESPONDENTSOLICITORONE;
 import static uk.gov.hmcts.reform.civil.enums.CaseRole.RESPONDENTSOLICITORTWO;
 import static uk.gov.hmcts.reform.civil.utils.ElementUtils.element;
@@ -123,39 +121,14 @@ abstract class EvidenceUploadHandlerBase extends CallbackHandler {
     protected static final String APPLICANT_TWO_WITNESS_HEARSAY = "ApplicantTwoWitnessHearsay";
     protected static final String RESPONDENT_ONE_WITNESS_HEARSAY = "RespondentOneWitnessHearsay";
     protected static final String RESPONDENT_TWO_WITNESS_HEARSAY = "RespondentTwoWitnessHearsay";
-    protected static final String RESPONDENT_ONE_DISCLOSURE_LIST = "RespondentOneDisclosureList";
-    protected static final String RESPONDENT_ONE_PRE_TRIAL_SUMMARY = "RespondentOnePreTrialSummary";
-    protected static final String RESPONDENT_ONE_TRIAL_SKELETON = "RespondentOneTrialSkeleton";
     protected static final String RESPONDENT_TWO_DISCLOSURE_LIST = "RespondentTwoDisclosureList";
     protected static final String RESPONDENT_TWO_PRE_TRIAL_SUMMARY = "RespondentTwoPreTrialSummary";
     protected static final String RESPONDENT_TWO_TRIAL_SKELETON = "RespondentTwoTrialSkeleton";
-    protected static final String RESPONDENT_ONE_SCHEDULE_OF_COSTS = "RespondentSchedulesOfCost";
     protected static final String RESPONDENT_TWO_SCHEDULE_OF_COSTS = "RespondentTwoSchedulesOfCost";
-    protected static final String APPLICANT_DISCLOSURE_LIST = "ApplicantDisclosureList";
-    protected static final String APPLICANT_PRE_TRIAL_SUMMARY = "ApplicantPreTrialSummary";
-    protected static final String APPLICANT_TRIAL_SKELETON = "ApplicantTrialSkeleton";
     protected static final String APPLICANT_TWO_DISCLOSURE_LIST = "ApplicantTwoDisclosureList";
     protected static final String APPLICANT_TWO_PRE_TRIAL_SUMMARY = "ApplicantTwoPreTrialSummary";
     protected static final String APPLICANT_TWO_TRIAL_SKELETON = "ApplicantTwoTrialSkeleton";
-    protected static final String APPLICANT_SCHEDULE_OF_COSTS = "ApplicantSchedulesOfCost";
     protected static final String APPLICANT_TWO_SCHEDULE_OF_COSTS = "ApplicantTwoSchedulesOfCost";
-    // Notification Strings used for email
-    protected static StringBuilder notificationString = new StringBuilder();
-    protected static final String DISCLOSURE_LIST_TEXT = "%s - Disclosure list";
-    protected static final String DISCLOSURE_TEXT = "%s - Documents for disclosure";
-    protected static final String WITNESS_STATEMENT_TEXT = "%s - Witness statement";
-    protected static final String WITNESS_SUMMARY_TEXT = "%s - Witness summary";
-    protected static final String WITNESS_HEARSAY_TEXT = "%s - Notice of the intention to rely on hearsay evidence";
-    protected static final String WITNESS_REFERRED_TEXT = "%s - Documents referred to in the statement";
-    protected static final String EXPERT_REPORT_TEXT = "%s - Expert's report";
-    protected static final String EXPERT_JOINT_STATEMENT_TEXT = "%s - Joint Statement of Experts / Single Joint Expert Report";
-    protected static final String EXPERT_QUESTIONS_TEXT = "%s - Questions for other party's expert or joint experts";
-    protected static final String EXPERT_ANSWERS_TEXT = "%s - Answer to questions asked";
-    protected static final String PRE_TRIAL_SUMMARY_TEXT = "%s - Case Summary";
-    protected static final String TRIAL_SKELETON_TEXT = "%s - Skeleton argument";
-    protected static final String TRIAL_AUTHORITIES_TEXT = "%s - Authorities";
-    protected static final String TRIAL_COSTS_TEXT = "%s - Costs";
-    protected static final String TRIAL_DOC_CORRESPONDENCE_TEXT = "%s - Documentary evidence for trial";
 
     protected static final String OPTION_APP1 = "Claimant 1 - ";
     protected static final String OPTION_APP2 = "Claimant 2 - ";
@@ -169,12 +142,15 @@ abstract class EvidenceUploadHandlerBase extends CallbackHandler {
     private static final String SELECTED_VALUE_APP_BOTH = "APPLICANTBOTH";
     private List<Element<UploadEvidenceDocumentType>> additionalBundleDocs;
 
+    private final SetOptionsTask setOptionsTask;
+    private final DocumentUploadTimeTask documentUploadTimeTask;
+
     @SuppressWarnings("java:S107")
     protected EvidenceUploadHandlerBase(UserService userService, CoreCaseUserService coreCaseUserService,
                                         CaseDetailsConverter caseDetailsConverter,
                                         CoreCaseDataService coreCaseDataService,
                                         ObjectMapper objectMapper, Time time, List<CaseEvent> events, String pageId,
-                                        String createShowCondition, FeatureToggleService featureToggleService) {
+                                        String createShowCondition, FeatureToggleService featureToggleService, SetOptionsTask setOptionsTask, DocumentUploadTimeTask documentUploadTimeTask) {
         this.objectMapper = objectMapper;
         this.time = time;
         this.createShowCondition = createShowCondition;
@@ -185,13 +161,13 @@ abstract class EvidenceUploadHandlerBase extends CallbackHandler {
         this.caseDetailsConverter = caseDetailsConverter;
         this.coreCaseDataService = coreCaseDataService;
         this.featureToggleService = featureToggleService;
+        this.setOptionsTask = setOptionsTask;
+        this.documentUploadTimeTask = documentUploadTimeTask;
     }
 
     abstract CallbackResponse validateValues(CallbackParams callbackParams, CaseData caseData);
 
     abstract CallbackResponse createShowCondition(CaseData caseData, UserInfo userInfo);
-
-    abstract void applyDocumentUploadDate(CaseData.CaseDataBuilder<?, ?> caseDataBuilder, LocalDateTime now);
 
     @Override
     public List<CaseEvent> handledEvents() {
@@ -210,40 +186,7 @@ abstract class EvidenceUploadHandlerBase extends CallbackHandler {
     }
 
     CallbackResponse setOptions(CallbackParams callbackParams) {
-        CaseData caseData = callbackParams.getCaseData();
-        List<String> dynamicListOptions = new ArrayList<>();
-        if (events.get(0).equals(EVIDENCE_UPLOAD_APPLICANT)) {
-            if (MultiPartyScenario.isTwoVOne(caseData)) {
-                dynamicListOptions.add(OPTION_APP1 + caseData.getApplicant1().getPartyName());
-                dynamicListOptions.add(OPTION_APP2 + caseData.getApplicant2().getPartyName());
-                dynamicListOptions.add(OPTION_APP_BOTH);
-            }
-        } else {
-            if (MultiPartyScenario.isOneVTwoLegalRep(caseData)) {
-                dynamicListOptions.add(OPTION_DEF1 + caseData.getRespondent1().getPartyName());
-                dynamicListOptions.add(OPTION_DEF2 + caseData.getRespondent2().getPartyName());
-                dynamicListOptions.add(OPTION_DEF_BOTH);
-            }
-        }
-        additionalBundleDocs = new ArrayList<>();
-        CaseData.CaseDataBuilder<?, ?> caseDataBuilder = caseData.toBuilder();
-        //Evidence upload will have different screen for Fast claims and Small claims.
-        // We use show hide in CCD to do this, using utility field caseProgAllocatedTrack to hold the value of the claim track
-        // for either spec claims (ResponseClaimTrack) or unspec claims (AllocatedTrack)
-        if (caseData.getCaseAccessCategory().equals(UNSPEC_CLAIM)) {
-            caseDataBuilder.caseProgAllocatedTrack(caseData.getAllocatedTrack().name());
-        } else if (caseData.getCaseAccessCategory().equals(SPEC_CLAIM)) {
-            caseDataBuilder.caseProgAllocatedTrack(caseData.getResponseClaimTrack());
-        }
-        caseDataBuilder.evidenceUploadOptions(DynamicList.fromList(dynamicListOptions));
-        // was unable to null value properly in EvidenceUploadNotificationEventHandler after emails are sent,
-        // so do it here if required.
-        if (nonNull(caseData.getNotificationText()) && caseData.getNotificationText().equals("NULLED")) {
-            caseDataBuilder.notificationText(null);
-        }
-        return AboutToStartOrSubmitCallbackResponse.builder()
-                .data(caseDataBuilder.build().toMap(objectMapper))
-                .build();
+        return setOptionsTask.setOptions(callbackParams.getCaseData());
     }
 
     CallbackResponse createShow(CallbackParams callbackParams) {
@@ -435,27 +378,6 @@ abstract class EvidenceUploadHandlerBase extends CallbackHandler {
         });
     }
 
-    public <T> void setCategoryIdAndRenameDoc(List<Element<T>> documentUpload, Function<Element<T>,
-        Document> documentExtractor, String theID, String docNotificationText, Function<Element<T>,
-        LocalDateTime> documentDateTimeExtractor, String claimantDefendantString) {
-        if (documentUpload == null || documentUpload.isEmpty()) {
-            return;
-        }
-        LocalDateTime halfFivePmYesterday = LocalDateTime.of(LocalDate.now().minusDays(1), LocalTime.of(17, 30));
-        renameDocuments(documentUpload, theID);
-        documentUpload.forEach(document -> {
-            Document documentToAddId = documentExtractor.apply(document);
-            documentToAddId.setCategoryID(theID);
-            LocalDateTime dateTime = documentDateTimeExtractor.apply(document);
-            if (dateTime.isAfter(halfFivePmYesterday)) {
-                String updateNotificationText = format(docNotificationText, claimantDefendantString);
-                if (!notificationString.toString().contains(updateNotificationText)) {
-                    notificationString.append("\n").append(updateNotificationText);
-                }
-            }
-        });
-    }
-
     private <T> void renameDocuments(List<Element<T>> documentUpload, String theId) {
         String prefix = null;
         switch (theId) {
@@ -609,505 +531,9 @@ abstract class EvidenceUploadHandlerBase extends CallbackHandler {
         });
     }
 
-    static void getNotificationText(CaseData caseData) {
-        notificationString = new StringBuilder();
-        if (caseData.getNotificationText() != null) {
-            notificationString = new StringBuilder(caseData.getNotificationText());
-        }
-    }
-
-    abstract void updateDocumentListUploadedAfterBundle(CaseData.CaseDataBuilder<?, ?> caseDataBuilder, CaseData caseData);
-
-    private void updateDocumentListUploadedAfterBundle(CaseData caseData, CaseData.CaseDataBuilder<?, ?> caseDataBuilder) {
-        if (nonNull(caseData.getCaseBundles()) && !caseData.getCaseBundles().isEmpty()) {
-            updateDocumentListUploadedAfterBundle(caseDataBuilder, caseData);
-        }
-    }
-
     CallbackResponse documentUploadTime(CallbackParams callbackParams) {
-        CaseData caseData = callbackParams.getCaseData();
-        CaseData.CaseDataBuilder<?, ?> caseDataBuilder = caseData.toBuilder();
-        // If notification has already been populated in current day, we want to append to that existing notification
-        getNotificationText(caseData);
-        applyDocumentUploadDate(caseDataBuilder, time.now());
-        updateDocumentListUploadedAfterBundle(caseData, caseDataBuilder);
-
         String selectedRole = getSelectedRole(callbackParams);
-        if (selectedRole.equals(RESPONDENTSOLICITORONE.name()) || selectedRole.equals(SELECTED_VALUE_DEF_BOTH)) {
-            String defendantString = "Defendant 1";
-            if (selectedRole.equals(SELECTED_VALUE_DEF_BOTH)) {
-                defendantString = "Both defendants";
-            }
-            setCategoryIdAndRenameDoc(caseData.getDocumentDisclosureListRes(), document -> document.getValue().getDocumentUpload(), RESPONDENT_ONE_DISCLOSURE_LIST,
-                                      DISCLOSURE_LIST_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentForDisclosureRes(), document -> document.getValue().getDocumentUpload(), RESPONDENT_ONE_DISCLOSURE,
-                                      DISCLOSURE_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentWitnessStatementRes(), document -> document.getValue().getWitnessOptionDocument(), RESPONDENT_ONE_WITNESS_STATEMENT,
-                                      WITNESS_STATEMENT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentWitnessSummaryRes(), document -> document.getValue().getWitnessOptionDocument(), RESPONDENT_ONE_WITNESS_SUMMARY,
-                                      WITNESS_SUMMARY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentHearsayNoticeRes(), document -> document.getValue().getWitnessOptionDocument(), RESPONDENT_ONE_WITNESS_HEARSAY,
-                                      WITNESS_HEARSAY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentReferredInStatementRes(), document -> document.getValue().getDocumentUpload(), RESPONDENT_ONE_WITNESS_REFERRED,
-                                      WITNESS_REFERRED_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentExpertReportRes(), document -> document.getValue().getExpertDocument(), RESPONDENT_ONE_EXPERT_REPORT,
-                                      EXPERT_REPORT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentJointStatementRes(), document -> document.getValue().getExpertDocument(), RESPONDENT_ONE_EXPERT_JOINT_STATEMENT,
-                                      EXPERT_JOINT_STATEMENT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentQuestionsRes(), document -> document.getValue().getExpertDocument(), RESPONDENT_ONE_EXPERT_QUESTIONS,
-                                      EXPERT_QUESTIONS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentAnswersRes(), document -> document.getValue().getExpertDocument(), RESPONDENT_ONE_EXPERT_ANSWERS,
-                                      EXPERT_ANSWERS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentCaseSummaryRes(), document -> document.getValue().getDocumentUpload(), RESPONDENT_ONE_PRE_TRIAL_SUMMARY,
-                                      PRE_TRIAL_SUMMARY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentSkeletonArgumentRes(), document -> document.getValue().getDocumentUpload(), RESPONDENT_ONE_TRIAL_SKELETON,
-                                      TRIAL_SKELETON_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentAuthoritiesRes(), document -> document.getValue().getDocumentUpload(), RESPONDENT_ONE_TRIAL_SKELETON,
-                                      TRIAL_AUTHORITIES_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentCostsRes(), document -> document.getValue().getDocumentUpload(), RESPONDENT_ONE_SCHEDULE_OF_COSTS,
-                                      TRIAL_COSTS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentEvidenceForTrialRes(), document -> document.getValue().getDocumentUpload(), RESPONDENT_ONE_TRIAL_DOC_CORRESPONDENCE,
-                                      TRIAL_DOC_CORRESPONDENCE_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            if (selectedRole.equals(SELECTED_VALUE_DEF_BOTH)) {
-                caseData = copyResp1ChangesToResp2(caseData, caseDataBuilder);
-            }
-        }
-        if (selectedRole.equals(RESPONDENTSOLICITORTWO.name())) {
-            String defendantString =  "Defendant 2";
-            setCategoryIdAndRenameDoc(caseData.getDocumentDisclosureListRes2(), document -> document.getValue().getDocumentUpload(), RESPONDENT_TWO_DISCLOSURE_LIST,
-                                      DISCLOSURE_LIST_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentForDisclosureRes2(), document -> document.getValue().getDocumentUpload(), RESPONDENT_TWO_DISCLOSURE,
-                                      DISCLOSURE_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentWitnessStatementRes2(), document -> document.getValue().getWitnessOptionDocument(), RESPONDENT_TWO_WITNESS_STATEMENT,
-                                      WITNESS_STATEMENT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentWitnessSummaryRes2(), document -> document.getValue().getWitnessOptionDocument(), RESPONDENT_TWO_WITNESS_SUMMARY,
-                                      WITNESS_SUMMARY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentHearsayNoticeRes2(), document -> document.getValue().getWitnessOptionDocument(), RESPONDENT_TWO_WITNESS_HEARSAY,
-                                      WITNESS_HEARSAY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentReferredInStatementRes2(), document -> document.getValue().getDocumentUpload(), RESPONDENT_TWO_WITNESS_REFERRED,
-                                      WITNESS_REFERRED_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentExpertReportRes2(), document -> document.getValue().getExpertDocument(), RESPONDENT_TWO_EXPERT_REPORT,
-                                      EXPERT_REPORT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentJointStatementRes2(), document -> document.getValue().getExpertDocument(), RESPONDENT_TWO_EXPERT_JOINT_STATEMENT,
-                                      EXPERT_JOINT_STATEMENT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentQuestionsRes2(), document -> document.getValue().getExpertDocument(), RESPONDENT_TWO_EXPERT_QUESTIONS,
-                                      EXPERT_QUESTIONS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentAnswersRes2(), document -> document.getValue().getExpertDocument(), RESPONDENT_TWO_EXPERT_ANSWERS,
-                                      EXPERT_ANSWERS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentCaseSummaryRes2(), document -> document.getValue().getDocumentUpload(), RESPONDENT_TWO_PRE_TRIAL_SUMMARY,
-                                      PRE_TRIAL_SUMMARY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentSkeletonArgumentRes2(), document -> document.getValue().getDocumentUpload(), RESPONDENT_TWO_TRIAL_SKELETON,
-                                      TRIAL_SKELETON_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentAuthoritiesRes2(), document -> document.getValue().getDocumentUpload(), RESPONDENT_TWO_TRIAL_SKELETON,
-                                      TRIAL_AUTHORITIES_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentCostsRes2(), document -> document.getValue().getDocumentUpload(), RESPONDENT_TWO_SCHEDULE_OF_COSTS,
-                                      TRIAL_COSTS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentEvidenceForTrialRes2(), document -> document.getValue().getDocumentUpload(), RESPONDENT_TWO_TRIAL_DOC_CORRESPONDENCE,
-                                      TRIAL_DOC_CORRESPONDENCE_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), defendantString);
-        }
-
-        if (selectedRole.equals(CaseRole.APPLICANTSOLICITORONE.name()) || selectedRole.equals(SELECTED_VALUE_APP_BOTH)) {
-            String claimantString = "Claimant 1";
-            if (selectedRole.equals(SELECTED_VALUE_APP_BOTH)) {
-                claimantString = "Both claimants";
-            }
-            setCategoryIdAndRenameDoc(caseData.getDocumentDisclosureList(), document -> document.getValue().getDocumentUpload(),  APPLICANT_DISCLOSURE_LIST,
-                                      DISCLOSURE_LIST_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentForDisclosure(), document -> document.getValue().getDocumentUpload(), APPLICANT_DISCLOSURE,
-                                      DISCLOSURE_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentWitnessStatement(), document -> document.getValue().getWitnessOptionDocument(), APPLICANT_WITNESS_STATEMENT,
-                                      WITNESS_STATEMENT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentWitnessSummary(), document -> document.getValue().getWitnessOptionDocument(), APPLICANT_WITNESS_SUMMARY,
-                                      WITNESS_SUMMARY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentHearsayNotice(), document -> document.getValue().getWitnessOptionDocument(), APPLICANT_WITNESS_HEARSAY,
-                                      WITNESS_HEARSAY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentReferredInStatement(), document -> document.getValue().getDocumentUpload(), APPLICANT_WITNESS_REFERRED,
-                                      WITNESS_REFERRED_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentExpertReport(), document -> document.getValue().getExpertDocument(), APPLICANT_EXPERT_REPORT,
-                                      EXPERT_REPORT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentJointStatement(), document -> document.getValue().getExpertDocument(), APPLICANT_EXPERT_JOINT_STATEMENT,
-                                      EXPERT_JOINT_STATEMENT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentQuestions(), document -> document.getValue().getExpertDocument(), APPLICANT_EXPERT_QUESTIONS,
-                                      EXPERT_QUESTIONS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentAnswers(), document -> document.getValue().getExpertDocument(), APPLICANT_EXPERT_ANSWERS,
-                                      EXPERT_ANSWERS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentCaseSummary(), document -> document.getValue().getDocumentUpload(), APPLICANT_PRE_TRIAL_SUMMARY,
-                                      PRE_TRIAL_SUMMARY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentSkeletonArgument(), document -> document.getValue().getDocumentUpload(), APPLICANT_TRIAL_SKELETON,
-                                      TRIAL_SKELETON_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentAuthorities(), document -> document.getValue().getDocumentUpload(), APPLICANT_TRIAL_SKELETON,
-                                      TRIAL_AUTHORITIES_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentCosts(), document -> document.getValue().getDocumentUpload(), APPLICANT_SCHEDULE_OF_COSTS,
-                                      TRIAL_COSTS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentEvidenceForTrial(), document -> document.getValue().getDocumentUpload(), APPLICANT_TRIAL_DOC_CORRESPONDENCE,
-                                      TRIAL_DOC_CORRESPONDENCE_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            if (selectedRole.equals(SELECTED_VALUE_APP_BOTH)) {
-                caseData = copyApp1ChangesToApp2(caseData, caseDataBuilder);
-            }
-        }
-
-        if (selectedRole.equals("APPLICANTSOLICITORTWO")) {
-            String claimantString =  "Claimant 2";
-            setCategoryIdAndRenameDoc(caseData.getDocumentDisclosureListApp2(), document -> document.getValue().getDocumentUpload(), APPLICANT_TWO_DISCLOSURE_LIST,
-                                      DISCLOSURE_LIST_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentForDisclosureApp2(), document -> document.getValue().getDocumentUpload(), APPLICANT_TWO_DISCLOSURE,
-                                      DISCLOSURE_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentWitnessStatementApp2(), document -> document.getValue().getWitnessOptionDocument(), APPLICANT_TWO_WITNESS_STATEMENT,
-                                      WITNESS_STATEMENT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentWitnessSummaryApp2(), document -> document.getValue().getWitnessOptionDocument(), APPLICANT_TWO_WITNESS_SUMMARY,
-                                      WITNESS_SUMMARY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentHearsayNoticeApp2(), document -> document.getValue().getWitnessOptionDocument(), APPLICANT_TWO_WITNESS_HEARSAY,
-                                      WITNESS_HEARSAY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentReferredInStatementApp2(), document -> document.getValue().getDocumentUpload(), APPLICANT_TWO_WITNESS_REFERRED,
-                                      WITNESS_REFERRED_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentExpertReportApp2(), document -> document.getValue().getExpertDocument(), APPLICANT_TWO_EXPERT_REPORT,
-                                      EXPERT_REPORT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentJointStatementApp2(), document -> document.getValue().getExpertDocument(), APPLICANT_TWO_EXPERT_JOINT_STATEMENT,
-                                      EXPERT_JOINT_STATEMENT_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentQuestionsApp2(), document -> document.getValue().getExpertDocument(), APPLICANT_TWO_EXPERT_QUESTIONS,
-                                      EXPERT_QUESTIONS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentAnswersApp2(), document -> document.getValue().getExpertDocument(), APPLICANT_TWO_EXPERT_ANSWERS,
-                                      EXPERT_ANSWERS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentCaseSummaryApp2(), document -> document.getValue().getDocumentUpload(), APPLICANT_TWO_PRE_TRIAL_SUMMARY,
-                                      PRE_TRIAL_SUMMARY_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentSkeletonArgumentApp2(), document -> document.getValue().getDocumentUpload(), APPLICANT_TWO_TRIAL_SKELETON,
-                                      TRIAL_SKELETON_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentAuthoritiesApp2(), document -> document.getValue().getDocumentUpload(), APPLICANT_TWO_TRIAL_SKELETON,
-                                      TRIAL_AUTHORITIES_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentCostsApp2(), document -> document.getValue().getDocumentUpload(), APPLICANT_TWO_SCHEDULE_OF_COSTS,
-                                      TRIAL_COSTS_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-            setCategoryIdAndRenameDoc(caseData.getDocumentEvidenceForTrialApp2(), document -> document.getValue().getDocumentUpload(), APPLICANT_TWO_TRIAL_DOC_CORRESPONDENCE,
-                                      TRIAL_DOC_CORRESPONDENCE_TEXT, documentDateTime -> documentDateTime.getValue().getCreatedDatetime(), claimantString);
-        }
-
-        // null the values of the lists, so that on future retriggers of the event, they are blank
-        caseDataBuilder.disclosureSelectionEvidence(null);
-        caseDataBuilder.disclosureSelectionEvidenceRes(null);
-        caseDataBuilder.witnessSelectionEvidence(null);
-        caseDataBuilder.witnessSelectionEvidenceSmallClaim(null);
-        caseDataBuilder.witnessSelectionEvidenceRes(null);
-        caseDataBuilder.witnessSelectionEvidenceSmallClaimRes(null);
-        caseDataBuilder.expertSelectionEvidenceRes(null);
-        caseDataBuilder.expertSelectionEvidence(null);
-        caseDataBuilder.expertSelectionEvidenceSmallClaim(null);
-        caseDataBuilder.expertSelectionEvidenceSmallClaimRes(null);
-        caseDataBuilder.trialSelectionEvidence(null);
-        caseDataBuilder.trialSelectionEvidenceSmallClaim(null);
-        caseDataBuilder.trialSelectionEvidenceRes(null);
-        caseDataBuilder.trialSelectionEvidenceSmallClaimRes(null);
-        caseDataBuilder.notificationText(notificationString.toString());
-
-        if (featureToggleService.isCaseProgressionEnabled()) {
-            caseDataBuilder.businessProcess(BusinessProcess.ready(EVIDENCE_UPLOADED));
-        }
-
-        return AboutToStartOrSubmitCallbackResponse.builder()
-            .data(caseDataBuilder.build().toMap(objectMapper))
-            .build();
-    }
-
-    private CaseData copyResp1ChangesToResp2(CaseData caseData, CaseData.CaseDataBuilder<?, ?> builder) {
-        CaseData caseDataBefore = caseDetailsConverter
-                .toCaseData(coreCaseDataService.getCase(caseData.getCcdCaseReference()));
-        List<Element<UploadEvidenceDocumentType>> evidenceDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentDisclosureListRes(),
-                    caseData.getDocumentDisclosureListRes(),
-                    caseData.getDocumentDisclosureListRes2());
-        List<Element<UploadEvidenceDocumentType>> evidenceDocToAdd =
-                deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, RESPONDENT_TWO_DISCLOSURE_LIST);
-        builder.documentDisclosureListRes2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentForDisclosureRes(),
-                                caseData.getDocumentForDisclosureRes(),
-                                caseData.getDocumentForDisclosureRes2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, RESPONDENT_TWO_DISCLOSURE);
-        builder.documentForDisclosureRes2(evidenceDocToAdd);
-
-        List<Element<UploadEvidenceWitness>> witnessDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentWitnessStatementRes(),
-                    caseData.getDocumentWitnessStatementRes(),
-                    caseData.getDocumentWitnessStatementRes2());
-        List<Element<UploadEvidenceWitness>> witnessDocToAdd =
-                deepCopyUploadEvidenceWitness(witnessDocToCopy, RESPONDENT_TWO_WITNESS_STATEMENT);
-        builder.documentWitnessStatementRes2(witnessDocToAdd);
-
-        witnessDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentWitnessSummaryRes(),
-                        caseData.getDocumentWitnessSummaryRes(),
-                        caseData.getDocumentWitnessSummaryRes2());
-        witnessDocToAdd =
-                deepCopyUploadEvidenceWitness(witnessDocToCopy, RESPONDENT_TWO_WITNESS_SUMMARY);
-        builder.documentWitnessSummaryRes2(witnessDocToAdd);
-
-        witnessDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentHearsayNoticeRes(),
-                        caseData.getDocumentHearsayNoticeRes(),
-                        caseData.getDocumentHearsayNoticeRes2());
-        witnessDocToAdd =
-                deepCopyUploadEvidenceWitness(witnessDocToCopy, RESPONDENT_TWO_WITNESS_HEARSAY);
-        builder.documentHearsayNoticeRes2(witnessDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentReferredInStatementRes(),
-                caseData.getDocumentReferredInStatementRes(),
-                caseData.getDocumentReferredInStatementRes2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, RESPONDENT_TWO_WITNESS_REFERRED);
-        builder.documentReferredInStatementRes2(evidenceDocToAdd);
-
-        List<Element<UploadEvidenceExpert>> expertDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentExpertReportRes(),
-                        caseData.getDocumentExpertReportRes(),
-                        caseData.getDocumentExpertReportRes2());
-        List<Element<UploadEvidenceExpert>> expertDocToAdd =
-                deepCopyUploadEvidenceExpert(expertDocToCopy, RESPONDENT_TWO_EXPERT_REPORT);
-        builder.documentExpertReportRes2(expertDocToAdd);
-
-        expertDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentJointStatementRes(),
-                        caseData.getDocumentJointStatementRes(),
-                        caseData.getDocumentJointStatementRes2());
-        expertDocToAdd =
-                deepCopyUploadEvidenceExpert(expertDocToCopy, RESPONDENT_TWO_EXPERT_JOINT_STATEMENT);
-        builder.documentJointStatementRes2(expertDocToAdd);
-
-        expertDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentQuestionsRes(),
-                        caseData.getDocumentQuestionsRes(),
-                        caseData.getDocumentQuestionsRes2());
-        expertDocToAdd =
-                deepCopyUploadEvidenceExpert(expertDocToCopy, RESPONDENT_TWO_EXPERT_QUESTIONS);
-        builder.documentQuestionsRes2(expertDocToAdd);
-
-        expertDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentAnswersRes(),
-                        caseData.getDocumentAnswersRes(),
-                        caseData.getDocumentAnswersRes2());
-        expertDocToAdd =
-                deepCopyUploadEvidenceExpert(expertDocToCopy, RESPONDENT_TWO_EXPERT_ANSWERS);
-        builder.documentAnswersRes2(expertDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentCaseSummaryRes(),
-                caseData.getDocumentCaseSummaryRes(),
-                caseData.getDocumentCaseSummaryRes2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, RESPONDENT_TWO_PRE_TRIAL_SUMMARY);
-        builder.documentCaseSummaryRes2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentSkeletonArgumentRes(),
-                caseData.getDocumentSkeletonArgumentRes(),
-                caseData.getDocumentSkeletonArgumentRes2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, RESPONDENT_TWO_TRIAL_SKELETON);
-        builder.documentSkeletonArgumentRes2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentAuthoritiesRes(),
-                caseData.getDocumentAuthoritiesRes(),
-                caseData.getDocumentAuthoritiesRes2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, RESPONDENT_TWO_TRIAL_SKELETON);
-        builder.documentAuthoritiesRes2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentCostsRes(),
-                caseData.getDocumentCostsRes(),
-                caseData.getDocumentCostsRes2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, RESPONDENT_TWO_SCHEDULE_OF_COSTS);
-        builder.documentCostsRes2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentEvidenceForTrialRes(),
-                caseData.getDocumentEvidenceForTrialRes(),
-                caseData.getDocumentEvidenceForTrialRes2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, RESPONDENT_TWO_TRIAL_DOC_CORRESPONDENCE);
-        builder.documentEvidenceForTrialRes2(evidenceDocToAdd);
-
-        return builder.build();
-    }
-
-    private CaseData copyApp1ChangesToApp2(CaseData caseData, CaseData.CaseDataBuilder<?, ?> builder) {
-        CaseData caseDataBefore = caseDetailsConverter
-                .toCaseData(coreCaseDataService.getCase(caseData.getCcdCaseReference()));
-        List<Element<UploadEvidenceDocumentType>> evidenceDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentDisclosureList(),
-                        caseData.getDocumentDisclosureList(),
-                        caseData.getDocumentDisclosureListApp2());
-        List<Element<UploadEvidenceDocumentType>> evidenceDocToAdd =
-                deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, APPLICANT_TWO_DISCLOSURE_LIST);
-        builder.documentDisclosureListApp2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentForDisclosure(),
-                caseData.getDocumentForDisclosure(),
-                caseData.getDocumentForDisclosureApp2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, APPLICANT_TWO_DISCLOSURE);
-        builder.documentForDisclosureApp2(evidenceDocToAdd);
-
-        List<Element<UploadEvidenceWitness>> witnessDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentWitnessStatement(),
-                        caseData.getDocumentWitnessStatement(),
-                        caseData.getDocumentWitnessStatementApp2());
-        List<Element<UploadEvidenceWitness>> witnessDocToAdd =
-                deepCopyUploadEvidenceWitness(witnessDocToCopy, APPLICANT_TWO_WITNESS_STATEMENT);
-        builder.documentWitnessStatementApp2(witnessDocToAdd);
-
-        witnessDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentWitnessSummary(),
-                        caseData.getDocumentWitnessSummary(),
-                        caseData.getDocumentWitnessSummaryApp2());
-        witnessDocToAdd =
-                deepCopyUploadEvidenceWitness(witnessDocToCopy, APPLICANT_TWO_WITNESS_SUMMARY);
-        builder.documentWitnessSummaryApp2(witnessDocToAdd);
-
-        witnessDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentHearsayNotice(),
-                        caseData.getDocumentHearsayNotice(),
-                        caseData.getDocumentHearsayNoticeApp2());
-        witnessDocToAdd =
-                deepCopyUploadEvidenceWitness(witnessDocToCopy, APPLICANT_TWO_WITNESS_HEARSAY);
-        builder.documentHearsayNoticeApp2(witnessDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentReferredInStatement(),
-                caseData.getDocumentReferredInStatement(),
-                caseData.getDocumentReferredInStatementApp2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, APPLICANT_TWO_WITNESS_REFERRED);
-        builder.documentReferredInStatementApp2(evidenceDocToAdd);
-
-        List<Element<UploadEvidenceExpert>> expertDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentExpertReport(),
-                        caseData.getDocumentExpertReport(),
-                        caseData.getDocumentExpertReportApp2());
-        List<Element<UploadEvidenceExpert>> expertDocToAdd =
-                deepCopyUploadEvidenceExpert(expertDocToCopy, APPLICANT_TWO_EXPERT_REPORT);
-        builder.documentExpertReportApp2(expertDocToAdd);
-
-        expertDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentJointStatement(),
-                        caseData.getDocumentJointStatement(),
-                        caseData.getDocumentJointStatementApp2());
-        expertDocToAdd =
-                deepCopyUploadEvidenceExpert(expertDocToCopy, APPLICANT_TWO_EXPERT_JOINT_STATEMENT);
-        builder.documentJointStatementApp2(expertDocToAdd);
-
-        expertDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentQuestions(),
-                        caseData.getDocumentQuestions(),
-                        caseData.getDocumentQuestionsApp2());
-        expertDocToAdd =
-                deepCopyUploadEvidenceExpert(expertDocToCopy, APPLICANT_TWO_EXPERT_QUESTIONS);
-        builder.documentQuestionsApp2(expertDocToAdd);
-
-        expertDocToCopy =
-                compareAndCopy(caseDataBefore.getDocumentAnswers(),
-                        caseData.getDocumentAnswers(),
-                        caseData.getDocumentAnswersApp2());
-        expertDocToAdd =
-                deepCopyUploadEvidenceExpert(expertDocToCopy, APPLICANT_TWO_EXPERT_ANSWERS);
-        builder.documentAnswersApp2(expertDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentCaseSummary(),
-                caseData.getDocumentCaseSummary(),
-                caseData.getDocumentCaseSummaryApp2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, APPLICANT_TWO_PRE_TRIAL_SUMMARY);
-        builder.documentCaseSummaryApp2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentSkeletonArgument(),
-                caseData.getDocumentSkeletonArgument(),
-                caseData.getDocumentSkeletonArgumentApp2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, APPLICANT_TWO_TRIAL_SKELETON);
-        builder.documentSkeletonArgumentApp2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentAuthorities(),
-                caseData.getDocumentAuthorities(),
-                caseData.getDocumentAuthoritiesApp2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, APPLICANT_TWO_TRIAL_SKELETON);
-        builder.documentAuthoritiesApp2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentCosts(),
-                caseData.getDocumentCosts(),
-                caseData.getDocumentCostsApp2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, APPLICANT_TWO_SCHEDULE_OF_COSTS);
-        builder.documentCostsApp2(evidenceDocToAdd);
-
-        evidenceDocToCopy = compareAndCopy(caseDataBefore.getDocumentEvidenceForTrial(),
-                caseData.getDocumentEvidenceForTrial(),
-                caseData.getDocumentEvidenceForTrialApp2());
-        evidenceDocToAdd = deepCopyUploadEvidenceDocumentType(evidenceDocToCopy, APPLICANT_TWO_TRIAL_DOC_CORRESPONDENCE);
-        builder.documentEvidenceForTrialApp2(evidenceDocToAdd);
-
-        return builder.build();
-    }
-
-    private List<Element<UploadEvidenceDocumentType>> deepCopyUploadEvidenceDocumentType(
-            final List<Element<UploadEvidenceDocumentType>> toCopy, String theId) {
-        if (Objects.isNull(toCopy)) {
-            return null;
-        }
-        List<Element<UploadEvidenceDocumentType>> toAdd = new ArrayList<>();
-        for (Element<UploadEvidenceDocumentType> from : toCopy) {
-            Document newDoc = Document.builder()
-                    .categoryID(theId)
-                    .documentBinaryUrl(from.getValue().getDocumentUpload().getDocumentBinaryUrl())
-                    .documentFileName(from.getValue().getDocumentUpload().getDocumentFileName())
-                    .documentHash(from.getValue().getDocumentUpload().getDocumentHash())
-                    .documentUrl(from.getValue().getDocumentUpload().getDocumentUrl())
-                    .build();
-            UploadEvidenceDocumentType type = UploadEvidenceDocumentType.builder()
-                    .witnessOptionName(from.getValue().getWitnessOptionName())
-                    .documentIssuedDate(from.getValue().getDocumentIssuedDate())
-                    .typeOfDocument(from.getValue().getTypeOfDocument())
-                    .createdDatetime(from.getValue().getCreatedDatetime())
-                    .documentUpload(newDoc)
-                    .build();
-            toAdd.add(ElementUtils.element(type));
-        }
-        return toAdd;
-    }
-
-    private List<Element<UploadEvidenceExpert>> deepCopyUploadEvidenceExpert(
-            final List<Element<UploadEvidenceExpert>> toCopy, String theId) {
-        if (Objects.isNull(toCopy)) {
-            return null;
-        }
-        List<Element<UploadEvidenceExpert>> toAdd = new ArrayList<>();
-        for (Element<UploadEvidenceExpert> from : toCopy) {
-            Document newDoc = Document.builder()
-                    .categoryID(theId)
-                    .documentBinaryUrl(from.getValue().getExpertDocument().getDocumentBinaryUrl())
-                    .documentFileName(from.getValue().getExpertDocument().getDocumentFileName())
-                    .documentHash(from.getValue().getExpertDocument().getDocumentHash())
-                    .documentUrl(from.getValue().getExpertDocument().getDocumentUrl())
-                    .build();
-            UploadEvidenceExpert type = UploadEvidenceExpert.builder()
-                    .expertOptionName(from.getValue().getExpertOptionName())
-                    .expertOptionExpertise(from.getValue().getExpertOptionExpertise())
-                    .expertOptionExpertises(from.getValue().getExpertOptionExpertises())
-                    .expertOptionOtherParty(from.getValue().getExpertOptionOtherParty())
-                    .expertDocumentQuestion(from.getValue().getExpertDocumentQuestion())
-                    .expertDocumentAnswer(from.getValue().getExpertDocumentAnswer())
-                    .expertOptionUploadDate(from.getValue().getExpertOptionUploadDate())
-                    .expertDocument(newDoc)
-                    .build();
-            toAdd.add(ElementUtils.element(type));
-        }
-        return toAdd;
-    }
-
-    private List<Element<UploadEvidenceWitness>> deepCopyUploadEvidenceWitness(
-            final List<Element<UploadEvidenceWitness>> toCopy, String theId) {
-        if (Objects.isNull(toCopy)) {
-            return null;
-        }
-        List<Element<UploadEvidenceWitness>> toAdd = new ArrayList<>();
-        for (Element<UploadEvidenceWitness> from : toCopy) {
-            Document newDoc = Document.builder()
-                    .categoryID(theId)
-                    .documentBinaryUrl(from.getValue().getWitnessOptionDocument().getDocumentBinaryUrl())
-                    .documentFileName(from.getValue().getWitnessOptionDocument().getDocumentFileName())
-                    .documentHash(from.getValue().getWitnessOptionDocument().getDocumentHash())
-                    .documentUrl(from.getValue().getWitnessOptionDocument().getDocumentUrl())
-                    .build();
-            UploadEvidenceWitness type = UploadEvidenceWitness.builder()
-                    .witnessOptionUploadDate(from.getValue().getWitnessOptionUploadDate())
-                    .witnessOptionName(from.getValue().getWitnessOptionName())
-                    .createdDatetime(from.getValue().getCreatedDatetime())
-                    .witnessOptionDocument(newDoc)
-                    .build();
-            toAdd.add(ElementUtils.element(type));
-        }
-        return toAdd;
+        return documentUploadTimeTask.documentUploadTime(callbackParams.getCaseData(), selectedRole);
     }
 
     protected static <T> List<Element<T>> compareAndCopy(List<Element<T>> before,
@@ -1164,7 +590,6 @@ abstract class EvidenceUploadHandlerBase extends CallbackHandler {
             return CaseRole.RESPONDENTSOLICITORONE.name();
         }
     }
-
     SubmittedCallbackResponse buildConfirmation(CallbackParams callbackParams) {
         return SubmittedCallbackResponse.builder()
             .confirmationHeader("# Documents uploaded")
